bot/requirements.txt
python-binance>=1.0.19
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.10
pydantic>=2.6.0
loguru>=0.7.2
python-telegram-bot>=20.7
pytest>=8.0.0
pydantic-settings>=2.2.1
